Quieter rebuild script for new builds.

+++ rebuild.sh
@@ -12,17 +12,28 @@ PROFILE="os_project"
 # Generate a complete Drupal build.
 echo "Building PROJECT..."
 
-# D7 has odd permissions on sites/default folder
-chmod 777 drupal/sites/default
 # remove old directories
-rm -rf drupal
-rm -rf $PROFILE/modules/contrib
-rm -rf $PROFILE/modules/stock
-rm -rf $PROFILE/themes/stock
-rm -rf $PROFILE/libraries
+if [ -d drupal ]; then
+  rm -rf drupal
+fi
+if [ -d $PROFILE/modules/contrib ]; then
+  rm -rf $PROFILE/modules/contrib
+fi
+if [ -d $PROFILE/modules/stock ]; then
+  rm -rf $PROFILE/modules/stock
+fi
+if [ -d $PROFILE/themes/stock ]; then
+  rm -rf $PROFILE/themes/stock
+fi
+if [ -d $PROFILE/libraries ]; then
+  rm -rf $PROFILE/libraries
+fi
 
 # Build the drupal directory from the make file.
 drush make $1 --yes --contrib-destination=profiles/$PROFILE $PROFILE/$PROFILE.make drupal
+
+# D7 has odd permissions on sites/default folder
+chmod 777 drupal/sites/default
 
 # Link settings.php and files into sites/default
 cd drupal/sites/default || { echo "Aborting $0"; exit 1; }
